Commited a large patch contributed by Ian Currington, which makes PoDoFo 64bit safe. He was able to edit a 5GB PDF file in one test.
Citing from his email:
Today, after several months of development,
we are able to contribute a new set of changes to PoDoFo.

These are specifically about 64-bit porting. As you can imagine, the
changes are widespread, with data type changes.
To enable you do perform difference analysis, the changes are against
SVN 990. The changes dramatically reduce cast and data type warnings
on MS Visual Studio C++ x64. The changes have been tested internally
here on 4 or 5 machines, including Linux 32-bit, Linux 64-bit, and
Vista 64, Visual Studio 2008 x64.



git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@1005 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ src/PdfAnnotation.cpp
@@ -134,7 +134,7 @@ bool PdfAnnotation::HasAppearanceStream() const
 
 void PdfAnnotation::SetFlags( pdf_uint32 uiFlags )
 {
-    m_pObject->GetDictionary().AddKey( "F", PdfVariant( static_cast<long>(uiFlags) ) );
+    m_pObject->GetDictionary().AddKey( "F", PdfVariant( static_cast<long long>(uiFlags) ) );
 }
 
 pdf_uint32 PdfAnnotation::GetFlags() const
